+++ Utilities/SilKitRegistry/Registry.cpp
@@ -513,7 +513,7 @@ int main(int argc, char** argv)
 
             if (useSignalHandler)
             {
-                using namespace SilKit::registry;
+                using namespace SilKit::Util;
 
                 auto signalValue = signalPromise.get_future();
                 RegisterSignalHandler([](auto sigNum) {
@@ -523,8 +523,7 @@ int main(int argc, char** argv)
 
                 signalValue.wait();
 
-                std::cout << "Signal " << signalValue.get() << " received!" << std::endl;
-                std::cout << "Exiting..." << std::endl;
+                std::cout << "Signal " << signalValue.get() << " received, exiting..." << std::endl;
             }
             else
             {